feat: refactor with component FicheReferenceCard
