+++ content/fiches-reference/index.ts
@@ -1,5 +1,6 @@
 // This file is generated by yarn build_fiches_reference. Run it again when fiches are added or removed.
 import { type FicheReference } from "@/lib/common/domain/values/FicheReference";
+
 import { ficheReference as ficheReference01_01 } from "./01_01";
 import { ficheReference as ficheReference01_02 } from "./01_02";
 import { ficheReference as ficheReference02_01 } from "./02_01";
@@ -7,20 +8,21 @@ import { ficheReference as ficheReference03_01 } from "./03_01";
 import { ficheReference as ficheReference04_01 } from "./04_01";
 import { ficheReference as ficheReference05_01 } from "./05_01";
 import { ficheReference as ficheReference06_01 } from "./06_01";
+import { ficheReference as ficheReference06_02 } from "./06_02";
 import { ficheReference as ficheReference07_01 } from "./07_01";
-import { ficheReference as ficheReference08_01 } from "./08_01";
 import { ficheReference as ficheReference10_01 } from "./10_01";
+import { ficheReference as ficheReference10_02 } from "./10_02";
+import { ficheReference as ficheReference10_03 } from "./10_03";
 import { ficheReference as ficheReference11_01 } from "./11_01";
 import { ficheReference as ficheReference12_01 } from "./12_01";
 import { ficheReference as ficheReference13_01 } from "./13_01";
+import { ficheReference as ficheReference13_02 } from "./13_02";
 import { ficheReference as ficheReference15_01 } from "./15_01";
 import { ficheReference as ficheReference16_01 } from "./16_01";
-import { ficheReference as ficheReference20_01 } from "./20_01";
 import { ficheReference as ficheReference21_01 } from "./21_01";
 import { ficheReference as ficheReference22_01 } from "./22_01";
 import { ficheReference as ficheReference23_01 } from "./23_01";
 import { ficheReference as ficheReference25_01 } from "./25_01";
-import { ficheReference as ficheReference26_01 } from "./26_01";
 import { ficheReference as ficheReference30_01 } from "./30_01";
 import { ficheReference as ficheReference31_01 } from "./31_01";
 import { ficheReference as ficheReference32_01 } from "./32_01";
@@ -41,20 +43,21 @@ export const fichesReference = {
   "04_01": ficheReference04_01,
   "05_01": ficheReference05_01,
   "06_01": ficheReference06_01,
+  "06_02": ficheReference06_02,
   "07_01": ficheReference07_01,
-  "08_01": ficheReference08_01,
   "10_01": ficheReference10_01,
+  "10_02": ficheReference10_02,
+  "10_03": ficheReference10_03,
   "11_01": ficheReference11_01,
   "12_01": ficheReference12_01,
   "13_01": ficheReference13_01,
+  "13_02": ficheReference13_02,
   "15_01": ficheReference15_01,
   "16_01": ficheReference16_01,
-  "20_01": ficheReference20_01,
   "21_01": ficheReference21_01,
   "22_01": ficheReference22_01,
   "23_01": ficheReference23_01,
   "25_01": ficheReference25_01,
-  "26_01": ficheReference26_01,
   "30_01": ficheReference30_01,
   "31_01": ficheReference31_01,
   "32_01": ficheReference32_01,
